Moved the logic for counting domain limits to a new class
